Solucionado tema de registrarse, me falta reacoplar el registro de empresa a lo de invitaciones tras relectura de requisitos. dejo pr igual

+++ src/ClassLibrary/LocationAPI/Location.cs
@@ -7,7 +7,7 @@
 namespace ClassLibrary.LocationAPI
 {
     /// <summary>
-    /// Representa las coordenadas y otros datos de la ubicación de una dirección retornada en el método
+    /// Representa las coordenadas y otros datos de la ubicación de una dirección retornada en el método.
     /// <see cref="LocationApiClient.GetLocationAsync(string, string, string, string)"/>.
     /// </summary>
     public class Location
@@ -53,20 +53,20 @@ namespace ClassLibrary.LocationAPI
         /// Obtiene o establece la latitud de la dirección.
         /// </summary>
         /// <value>El valor de la latitud en formato decimal.</value>
-        public double Latitude { get; set; }
+        public float Latitude { get; set; }
 
         /// <summary>
         /// Obtiene o establece la longitud de la dirección.
         /// </summary>
         /// <value>El valor de la longitud en formato decimal.</value>
-        public double Longitude { get; set; }
+        public float Longitude { get; set; }
 
-        public static bool operator ==(Location a, Location b)
+        public static bool operator == (Location a, Location b)
         {
             return a.Latitude == b.Latitude && a.Longitude == b.Longitude;
         }
 
-        public static bool operator !=(Location a, Location b)
+        public static bool operator != (Location a, Location b)
         {
             return a.Latitude != b.Latitude || a.Longitude != b.Longitude;
         }
